refactor: remediate open SonarCloud findings
Fixes 60 findings in code; documents scoped carve-outs for the 39 that fire on
correct-by-design code. No behavior change anywhere.

Fixed:
- tcp.cpp S3519 (blocker) was a false positive, not an overflow: the conn_pool
  write was already guarded by slot < CONN_POOL_SLOTS, but the guard sat inside
  an enclosing slot >= MAX_CONNS branch the analyzer could not reconcile. Bound
  first, then dispatch. That exposed something real - with HTTP/3 off the two
  are equal, so the reserved-slot branch is dead code the compiler warns on; it
  is now #if DWS_INTERNAL_SLOTS > 0, which is what it always meant.
- const-correctness (S995, S5350): dbm/mtconnect take pointer-to-const, their
  probe locals are pointer-to-const, failsafe's lifeline is reference-to-const.
  No const_cast; the mutating paths stay non-const.
- scpi (S3776 34, S886): parse_number split into scan_digits / scan_exponent /
  apply_scale, and the quote-unescape loop no longer advances its counter in two
  places.
- smtp (S3776 57 -> ~12): smtp_run split into six phase helpers.
- simatic (S3776 35, S134): 3964R parse_block lifts out put_byte / bcc_ok;
  measured nesting depth 6 -> 3.
- ssh_transport (S134): cand_match extracted from negotiate_alg, 4 -> 3.
  RFC 4253 s7.1 client-preference order untouched.
- sse merged nested ifs, forwarded_trust unnested a ternary, base64 moved a
  mid-file include to the top, dwserver dropped a redundant cast, and three
  comments were reworded so they stop tripping the commented-out-code heuristic.
- cyclone interop scripts: positional params bound to locals, explicit returns,
  [[ ]] tests, SUDO -> sudo_run.

Carved out in sonar-project.properties, each with the measurement behind it:
- sntrup761.cpp is vendored - OpenSSH's sntrup761, which itself vendors supercop
  djbsort. 34 of its 39 findings sit on verbatim upstream lines (crypto_sort_int32,
  R3_recip/Rq_recip3, 22 of 26 multi-declarators). This file is re-audited by
  diffing against upstream, so inherited lines keep upstream's shape; the file
  header now says so. The 4 project-written declarations and 2 redundant casts
  are fixed, not waived.
- S5416: 76 of src/'s 89 typedefs are C function-pointer callbacks and Sonar sees
  only 12 through the compile DB, so converting those 12 would split one uniform
  convention in two. Same plain-C stance already documented for S5205/S5008.
- S8417: the lock-free seams pick their orders deliberately and document the
  pairing at each site; seq_cst would put a full barrier on the accept path this
  replaced a linear scan to speed up.
- S959 / S954: derived_sizing.h exists to raise floors defined elsewhere, and
  ServerConfig.h must include it last or the floors evaluate against undefined
  macros. Genuine mid-file includes elsewhere are fixed, not waived.
- S923: a printf-style logger has to take varargs to reach vsnprintf.

Verified host-native in WSL: 22 envs, 780 test cases, 0 failures. The SCPI number
parser changes control flow in a float parser, so it was additionally tested
differentially old-vs-new over 66,430 inputs comparing both the bool result and
the exact double bits - 0 mismatches. clang-format 22.1.8 clean.

Not hardware-run: this is a host-verified refactor, not a HW-verified change.

Unrelated pre-existing issue found while verifying, left alone: native_pqc's
test_pqc_mlkem and test_pqc_sha3 do not link (sntrup761.cpp is in the env-wide
build_src_filter and references ssh_rng_fill, which only test_pqc_sntrup761.cpp
stubs). Reproduced at unmodified HEAD. Fixing it means editing test_matrix.json
and regenerating platformio.ini.

+++ src/dwserver.cpp
@@ -1207,7 +1207,7 @@ void DWS::http_poll_slot(uint8_t i)
     // already returned above.
     if (conn_pool[i].state == ConnState::CONN_ACTIVE && conn_pool[i].req_start_ms != 0 &&
         http_pool[i].parse_state < ParseState::PARSE_BODY &&
-        (uint32_t)(dws_millis() - conn_pool[i].req_start_ms) >= DWS_REQUEST_TIMEOUT_MS)
+        (dws_millis() - conn_pool[i].req_start_ms) >= DWS_REQUEST_TIMEOUT_MS)
     {
         conn_pool[i].req_start_ms = 0;
         send(i, 408, DWS_MIME_TEXT_PLAIN, "Request Timeout"); // terminal error response -> connection closes
